Change type of return on digitalRead function

+++ _arduino/wiring_digital.c
@@ -161,7 +161,7 @@ void digitalWrite(uint8_t pin, uint8_t val)
 	SREG = oldSREG;
 }
 
-uint8_t digitalRead(uint8_t pin)
+int digitalRead(uint8_t pin)
 {
 	uint8_t timer = digitalPinToTimer(pin);
 	uint8_t bit = digitalPinToBitMask(pin);
